Add uncompletion logic for single shared completion todos and dailies

+++ website/server/libs/groupTasks.js
@@ -8,23 +8,44 @@ const SHARED_COMPLETION = {
   every: 'allAssignedCompletion',
 };
 
-async function _completeMasterTask (masterTask) {
-  masterTask.completed = true;
+async function _completeOrUncompleteMasterTask (masterTask, completed) {
+  masterTask.completed = completed;
   await masterTask.save();
 }
 
-async function _deleteOrCompleteUnfinishedTasks (groupMemberTask) {
+async function _updateAssignedUsersTasks (masterTask, groupMemberTask) {
   if (groupMemberTask.type == 'todo') {
-    // The task was done by one person and is removed from others' lists
-    await Tasks.Task.deleteMany({
-      'group.taskId': groupMemberTask.group.taskId,
-      $and: [
-        {userId: {$exists: true}},
-        {userId: {$ne: groupMemberTask.userId}},
-      ],
-    }).exec();
+    if (groupMemberTask.completed) {
+      // The task was done by one person and is removed from others' lists
+      await Tasks.Task.deleteMany({
+        'group.taskId': groupMemberTask.group.taskId,
+        $and: [
+          {userId: {$exists: true}},
+          {userId: {$ne: groupMemberTask.userId}},
+        ],
+      }).exec();
+    } else {
+      // The task was uncompleted by the group member and should be recreated for assignedUsers
+      let group = await Groups.findById(masterTask.group.id);
+      let userList = [];
+      masterTask.group.assignedUsers.forEach(userId => {
+        let query = {_id: userId};
+        userList.push(query);
+      });
+      // @REVIEW There has to be a better way to do this
+      // async arrow function callback to the find and/or a forEach resulted in Mongoose ParallelSaveErrors
+      let assignedUsers = await Users.find({
+          $or: userList
+        }).exec();
+      for (let i=0; i < assignedUsers.length; i++) {
+        let promises = [];
+        promises.push(group.syncTask(masterTask, assignedUsers[i]));
+        promises.push(group.save());
+        await Promise.all(promises);
+      }
+    }
   } else {
-    // Complete the task on other users' lists
+    // Complete or uncomplete the task on other users' lists
     await Tasks.Task.find({
       'group.taskId': groupMemberTask.group.taskId,
       $and: [
@@ -36,12 +57,13 @@ async function _deleteOrCompleteUnfinishedTasks (groupMemberTask) {
         if (err) return;
 
         tasks.forEach (task => {
+          // Ajdust the task's completion to match the groupMemberTask
           // @REVIEW Completed or notDue tasks have no effect at cron
           // This maintain's the user's streak without scoring the task if someone else completed the task
           // If no assignedUser completes the due daily, all users lose their streaks at their cron
           // An alternative is to set the other assignedUsers' tasks to a later startDate
           // Should we break their streaks to encourage competition for the daily?
-          task.completed = true;
+          task.completed = groupMemberTask.completed;
           task.save();
         });
       });
@@ -62,9 +84,7 @@ async function _evaluateAllAssignedCompletion (masterTask) {
       completed: true,
     }).exec();
   }
-  if (completions >= masterTask.group.assignedUsers.length) {
-    await _completeMasterTask(masterTask);
-  }
+  await _completeOrUncompleteMasterTask(masterTask, (completions >= masterTask.group.assignedUsers.length));
 }
 
 async function handleSharedCompletion (groupMemberTask) {
@@ -75,8 +95,8 @@ async function handleSharedCompletion (groupMemberTask) {
   if (!masterTask || !masterTask.group || masterTask.type == 'habit') return;
 
   if (masterTask.group.sharedCompletion === SHARED_COMPLETION.single) {
-    await _deleteOrCompleteUnfinishedTasks(groupMemberTask);
-    await _completeMasterTask(masterTask);
+    await _updateAssignedUsersTasks(masterTask, groupMemberTask);
+    await _completeOrUncompleteMasterTask(masterTask, groupMemberTask.completed);
   } else if (masterTask.group.sharedCompletion === SHARED_COMPLETION.every) {
     await _evaluateAllAssignedCompletion(masterTask);
   }
